🚚 rename: ModalInfo를 organisms로 이동 #58
