+++ src/pages/MyTicket/components/modal/InformationModal/index.tsx
@@ -5,7 +5,7 @@ import ErrorPage from "@/pages/ErrorPage";
 import CancellationConfirmModal from "@/pages/MyTicket/components/modal/CancellationConfirmModal";
 import useHistoryModal from "@/hooks/useHistoryModal";
 import DetailedTicket from "@/common/components/molecules/DetailedTicket";
-import ModalInfo from "@/common/components/molecules/ModalInfo";
+import ModalInfo from "@/common/components/organisms/ModalInfo";
 import ModalHeader from "@/common/components/molecules/ModalHeader";
 import useReservationDetails from "@/pages/MyTicket/api/useReservationDetails";
